updated the data type of motor_tte_time in data collator

+++ src/cehrgpt/data/hf_cehrgpt_dataset_collator.py
@@ -545,8 +545,8 @@ class CehrGptDataCollator:
                     n_tte_predictions,
                     self.tokenizer.motor_tte_vocab_size,
                 ),
-                fill_value=0,
-                dtype=np.int32,
+                fill_value=0.0,
+                dtype=np.float32,
             )
             motor_tte_event_indicator = np.zeros(
                 (
